requirements.txt
# Core
alpaca-py>=0.30.0
yfinance>=0.2.0
edgartools>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
asyncpg>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Market calendar
exchange_calendars>=4.5.0

# LLM
anthropic>=0.40.0
groq>=0.11.0

# Dashboard (phase 9)
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
